+++ R/geom-measures.R
@@ -211,3 +211,52 @@ st_distance = function(x, y, ..., dist_fun, by_element = FALSE,
 		d
 	}
 }
+
+check_lengths = function (dots) {
+	lengths <- vapply(dots, length, integer(1))
+	non_constant_lengths <- unique(lengths[lengths != 1])
+	if (length(non_constant_lengths) == 0) {
+		1
+	}
+	else if (length(non_constant_lengths) == 1) {
+		non_constant_lengths
+	}
+	else {
+		lengths_label <- paste0(non_constant_lengths, collapse = ", ")
+		stop(sprintf("Incompatible lengths: %s", lengths_label), 
+			call. = FALSE)
+	}
+}
+
+recycle_common = function (dots) {
+	final_length <- check_lengths(dots)
+	lapply(dots, rep_len, final_length)
+}
+
+
+#' Project point on linestring, interpolate along a linestring
+#'
+#' Project point on linestring, interpolate along a linestring
+#' @param line object of class `sfc` with `LINESTRING` geometry
+#' @param point object of class `sfc` with `POINT` geometry
+#' @param normalized logical; if `TRUE`, use or return distance normalised to 0-1
+#' @name st_line_project_point
+#' @returns `st_line_project` returns the distance(s) of point(s) along line(s), when projected on the line(s)
+#' @export
+#' @details
+#' arguments `line`, `point` and `dist` are recycled to common length when needed
+#' @examples
+#' st_line_project(st_as_sfc("LINESTRING (0 0, 10 10)"), st_as_sfc(c("POINT (0 0)", "POINT (5 5)")))
+#' st_line_project(st_as_sfc("LINESTRING (0 0, 10 10)"), st_as_sfc("POINT (5 5)"), TRUE)
+st_line_project = function(line, point, normalized = FALSE) {
+	stopifnot(inherits(line, "sfc"), inherits(point, "sfc"),
+		all(st_dimension(line) == 1), all(st_dimension(point) == 0),
+		is.logical(normalized), length(normalized) == 1,
+		st_crs(line) == st_crs(point))
+	line = st_cast(line, "LINESTRING")
+	point = st_cast(point, "POINT")
+	if (isTRUE(st_is_longlat(line)))
+		message_longlat("st_project_point")
+	recycled = recycle_common(list(line, point))
+	CPL_line_project(recycled[[1]], recycled[[2]], normalized)
+}